requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
Jinja2==2.11.3
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-allauth==0.44.0
drf-spectacular==0.15.1